Merging bro analysis scripts


Former-commit-id: d7ef7955feb6e83a23814861fe1539504c8c40d5
Former-commit-id: 8bc8bdabdf7bc21568e1ff947461e344a3707d21
Former-commit-id: 038cbc8067d87902cb9e2d242306315410ca42b4
Former-commit-id: 6c2bd1cc7633bc5959fb00c0bfc220a1129ac088

+++ code/PcapProcessing/bro-analysis/run-bro.sh
@@ -1,41 +1,50 @@
 #!/bin/bash
 
-baseDir="/Users/ashwin/proj-work/meddle/meddle-data"
-currDir=${baseDir}
-decryptDir=${currDir}/pcap-decrypted
-broDir=${currDir}/bro-results/
+if [ $# -ne 1 ]
+then
+  echo "$# $0 <userName>"
+  exit -1
+fi
+reqUserName=$1
+baseDir="/user/arao/home/meddle_data/"
+broHome="/user/arao/home/bro/"
+broBin="${broHome}/bin/bro"
+broCustom="${broHome}/custom-scripts/"
+srcDir="${baseDir}/parsing-scripts/"
+decryptDir="${baseDir}/pcap-decrypted/"
+broDir="${baseDir}/bro-results/"
 mkdir -p ${broDir}
-tmpDir=${currDir}/tmpLogs/
+tmpDir="${baseDir}/tmpLogs/tmp-${reqUserName}/"
 mkdir -p ${tmpDir}
-srcDir="/Users/ashwin/proj-work/meddle/ashwin-meddle/meddle/code/PcapProcessing/bro-analysis/gen-analysis-logs/"
 for uDir in ${decryptDir}/*;
 do
     uName=`basename ${uDir}`
-    echo ${uName}
-    #if [ "${uName}" != "will-ipad" ];
-    #then
-    #    continue
-    #fi 
+    if [ "${uName}" != "${reqUserName}" ];
+    then
+        echo "${uName} != ${reqUserName}"
+        continue
+    fi 
+    echo "Running Bro for ${uName}"
     broUDir=${broDir}/${uName}
     mkdir -p ${broUDir}
     rm -f ${tmpDir}/*
-    cp -vRf ${currDir}/bro-headers/* ${broUDir}
+    cp -vRf ${baseDir}/bro-headers/* ${broUDir}
 
-    cnt=1
-    totCnt=1
+    cnt=0
+    totCnt=0
     for fName in ${uDir}/*.pcap*;
     do
-	cd ${tmpDir}
+        cd ${tmpDir}
         rm -f ${tmpDir}/*
-	bro -r ${fName} > /dev/null 2>&1
+        ${broBin} -r ${fName} ${broCustom}/*.bro  > /dev/null 2>&1
         globalIP=`python ${srcDir}/getIPFromFile.py ${fName} 1`
         localIP=`python ${srcDir}/getIPFromFile.py ${fName} 2`
-	#echo ${globalIP} ${localIP}
+        #echo ${globalIP} ${localIP}
         for logFile in *.log;
-	do
+        do
             sed -e s/"${localIP}"/"${globalIP}"/g ${logFile} > ${logFile}.tmp
             cat ${logFile}.tmp >> ${broUDir}/${logFile}
-	done
+        done
         # Will not work for traceroutes/sshs to localmachines, etc .. will not work for x.10.11.x
         #cnt=`grep -c "10\.11\." ${broUDir}/conn.log`
         #echo ${cnt}
@@ -46,21 +55,25 @@ do
         #   exit 
         #fi
         cnt=$((cnt+1))
-        if [ $cnt -gt 25 ] ;
+        if [ $cnt -eq 25 ] ;
         then
             totCnt=$((totCnt+cnt))
             echo $totCnt
-            cnt=1
+            cnt=0
         fi
-	cd - > /dev/null 2>&1
+        cd - > /dev/null 2>&1
     done
     cd ${broUDir}
     for logFile in *.log;
     do
-	sed '/^#/d' ${logFile} > ${uName}.log
-	mv ${uName}.log ${logFile}
+        sed '/^#/d' ${logFile} > ${uName}.log
+        mv ${uName}.log ${logFile}
     done
-    cd - > /dev/null 2>&1	
+    cd - > /dev/null 2>&1        
     rm -Rf ${tmpDir}/*
+    rm -Rf ${tmpDir}/.state
     #rmdir ${tmpDir}
 done
+rm -Rf ${tmpDir}/.state
+rmdir ${tmpDir}
+
